src/app/shared/components/product/camfil-product-rating-star/camfil-product-rating-star.component.spec.ts
import { ComponentFixture, TestBed, async } from '@angular/core/testing';
import { FaIconComponent } from '@fortawesome/angular-fontawesome';
import { MockComponent } from 'ng-mocks';

import { CamfilProductRatingStarComponent } from './camfil-product-rating-star.component';

describe('Product Rating Star Component', () => {
  let component: CamfilProductRatingStarComponent;
  let fixture: ComponentFixture<CamfilProductRatingStarComponent>;
  let element: HTMLElement;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [MockComponent(FaIconComponent), CamfilProductRatingStarComponent],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CamfilProductRatingStarComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
